requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
configparser==5.0.0
docker-pycreds==0.4.0
gitdb==4.0.4
GitPython==3.1.1
gql==0.2.0
graphql-core==1.1
idna==2.9
numpy==1.18.2
nvidia-ml-py3==7.352.0
pathtools==0.1.2
Pillow==7.1.1
promise==2.3
psutil==5.7.0
python-dateutil==2.8.1
PyYAML==5.3.1
requests==2.23.0
sentry-sdk==0.14.3
shortuuid==1.0.1
six==1.14.0
smmap==3.0.2
subprocess32==3.5.4
torch==1.4.0
torchvision==0.5.0
tqdm==4.45.0
urllib3==1.25.9
wandb==0.8.32
watchdog==0.10.2
